+++ IQKeyboardManagerSwift/IQKeyboardManager/UIKitExtensions/UIView+Parent.swift
@@ -22,6 +22,7 @@
 //  THE SOFTWARE.
 
 import UIKit
+import IQKeyboardCore
 
 /**
 UIView hierarchy category.
