+++ IQKeyboardManagerSwift/IQKeyboardNotification/IQKeyboardListener.swift
@@ -25,6 +25,9 @@ import UIKit
 import Combine
 
 @available(iOSApplicationExtension, unavailable)
+// swiftlint:disable line_length
+@available(*, deprecated, message: "Please use `IQKeyboardNotification` independently from https://github.com/hackiftekhar/IQKeyboardNotification. IQKeyboardListener will be removed from this library in future release.")
+// swiftlint:enable line_length
 @MainActor
 @objc public final class IQKeyboardListener: NSObject {
 
@@ -32,9 +35,9 @@ import Combine
 
     private var eventObservers: [AnyHashable: SizeCompletion] = [:]
 
-    public private(set) var oldKeyboardInfo: IQKeyboardInfo
+    public private(set) var oldKeyboardInfo: IQKeyboardInfoDeprecated
 
-    public private(set) var keyboardInfo: IQKeyboardInfo {
+    public private(set) var keyboardInfo: IQKeyboardInfoDeprecated {
         didSet {
             guard keyboardInfo != oldValue else { return }
             oldKeyboardInfo = oldValue
@@ -51,15 +54,15 @@ import Combine
     }
 
     @objc public override init() {
-        keyboardInfo = IQKeyboardInfo(notification: nil, event: .didHide)
+        keyboardInfo = IQKeyboardInfoDeprecated(notification: nil, event: .didHide)
         oldKeyboardInfo = keyboardInfo
 
         super.init()
 
         //  Registering for keyboard notification.
-        for event in IQKeyboardInfo.Event.allCases {
+        for event in IQKeyboardInfoDeprecated.Event.allCases {
             NotificationCenter.default.publisher(for: event.notification)
-                .map({ IQKeyboardInfo(notification: $0, event: event) })
+                .map({ IQKeyboardInfoDeprecated(notification: $0, event: event) })
                 .assign(to: \.keyboardInfo, on: self)
                 .store(in: &storage)
         }
@@ -68,27 +71,22 @@ import Combine
     @objc public func animate(alongsideTransition transition: @escaping () -> Void, completion: (() -> Void)? = nil) {
         keyboardInfo.animate(alongsideTransition: transition, completion: completion)
     }
-}
 
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-public extension IQKeyboardListener {
+    public typealias SizeCompletion = (_ event: IQKeyboardInfoDeprecated.Event, _ endFrame: CGRect) -> Void
 
-    typealias SizeCompletion = (_ event: IQKeyboardInfo.Event, _ endFrame: CGRect) -> Void
-
-    @objc func subscribe(identifier: AnyHashable, changeHandler: @escaping SizeCompletion) {
+    @objc public func subscribe(identifier: AnyHashable, changeHandler: @escaping SizeCompletion) {
         eventObservers[identifier] = changeHandler
     }
 
-    @objc func unsubscribe(identifier: AnyHashable) {
+    @objc public func unsubscribe(identifier: AnyHashable) {
         eventObservers[identifier] = nil
     }
 
-    @objc func isSubscribed(identifier: AnyHashable) -> Bool {
+    @objc public func isSubscribed(identifier: AnyHashable) -> Bool {
         eventObservers[identifier] != nil
     }
 
-    private func sendKeyboardInfo(info: IQKeyboardInfo) {
+    private func sendKeyboardInfo(info: IQKeyboardInfoDeprecated) {
 
         let endFrame: CGRect = info.endFrame
 
@@ -96,19 +94,11 @@ public extension IQKeyboardListener {
             block(info.event, endFrame)
         }
     }
-}
 
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-@objc public extension IQKeyboardListener {
+    // MARK: Deprecated
 
     @available(*, deprecated, renamed: "isVisible")
     var keyboardShowing: Bool { isVisible }
-}
-
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-public extension IQKeyboardListener {
 
     @available(*, deprecated, renamed: "subscribe(identifier:changeHandler:)")
     @objc func registerSizeChange(identifier: AnyHashable, changeHandler: @escaping SizeCompletion) {
