+++ IQKeyboardManagerSwift/IQKeyboardReturnKeyHandler/IQKeyboardReturnKeyHandler.swift
@@ -22,13 +22,20 @@
 //  THE SOFTWARE.
 
 import UIKit
+import IQKeyboardCore
 
+// swiftlint:disable file_length
 /**
 Manages the return key to work like next/done in a view hierarchy.
 */
+
 @available(iOSApplicationExtension, unavailable)
+// swiftlint:disable line_length
+@available(*, deprecated, message: "Please use `IQKeyboardReturnManager` independently from https://github.com/hackiftekhar/IQKeyboardReturnManager. IQKeyboardReturnKeyHandler will be removed from this library in future release.")
+// swiftlint:enable line_length
 @MainActor
-@objc public final class IQKeyboardReturnKeyHandler: NSObject {
+// swiftlint:disable type_body_length
+@objc public final class IQKeyboardReturnKeyHandler: NSObject, UITextFieldDelegate, UITextViewDelegate {
 
     // MARK: Private variables
     private var textInputViewInfoCache: [IQTextInputViewInfoModel] = []
@@ -83,20 +90,15 @@ Manages the return key to work like next/done in a view hierarchy.
 
         textInputViewInfoCache.removeAll()
     }
-}
 
-// MARK: Registering/Unregistering textInputView
-
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-public extension IQKeyboardReturnKeyHandler {
+    // MARK: Registering/Unregistering textInputView
 
     /**
      Should pass TextInputView instance. Assign textInputView delegate to self, change it's returnKeyType.
 
      @param view TextInputView object to register.
      */
-    @objc func add(textInputView: any IQTextInputView) {
+    @objc public func add(textInputView: any IQTextInputView) {
 
         let model = IQTextInputViewInfoModel(textInputView: textInputView)
         textInputViewInfoCache.append(model)
@@ -113,7 +115,7 @@ public extension IQKeyboardReturnKeyHandler {
 
      @param view TextInputView object to unregister.
      */
-    @objc func remove(textInputView: any IQTextInputView) {
+    @objc public func remove(textInputView: any IQTextInputView) {
 
         guard let index: Int = textInputViewCachedInfoIndex(textInputView) else { return }
 
@@ -122,11 +124,11 @@ public extension IQKeyboardReturnKeyHandler {
     }
 
     /**
-    Add all the TextInputView responderView's.
+     Add all the TextInputView responderView's.
 
-    @param view object to register all it's responder subviews.
-    */
-    @objc func addResponderSubviews(of view: UIView, recursive: Bool) {
+     @param view object to register all it's responder subviews.
+     */
+    @objc public func addResponderSubviews(of view: UIView, recursive: Bool) {
 
         let textInputViews: [any IQTextInputView] = view.responderSubviews(recursive: recursive)
 
@@ -136,11 +138,11 @@ public extension IQKeyboardReturnKeyHandler {
     }
 
     /**
-    Remove all the TextInputView responderView's.
+     Remove all the TextInputView responderView's.
 
-    @param view object to unregister all it's responder subviews.
-    */
-    @objc func removeResponderSubviews(of view: UIView, recursive: Bool) {
+     @param view object to unregister all it's responder subviews.
+     */
+    @objc public func removeResponderSubviews(of view: UIView, recursive: Bool) {
 
         let textInputViews: [any IQTextInputView] = view.responderSubviews(recursive: recursive)
 
@@ -148,13 +150,9 @@ public extension IQKeyboardReturnKeyHandler {
             remove(textInputView: view)
         }
     }
-}
 
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-public extension IQKeyboardReturnKeyHandler {
     @discardableResult
-    func goToNextResponderOrResign(from textInputView: some IQTextInputView) -> Bool {
+    public func goToNextResponderOrResign(from textInputView: some IQTextInputView) -> Bool {
 
         guard let textInfoCache: IQTextInputViewInfoModel = nextResponderFromTextInputView(textInputView),
               let textInputView = textInfoCache.textInputView else {
@@ -165,16 +163,12 @@ public extension IQKeyboardReturnKeyHandler {
         textInputView.becomeFirstResponder()
         return false
     }
-}
 
-// MARK: Internal Functions
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-internal extension IQKeyboardReturnKeyHandler {
+    // MARK: Internal Functions
 
-    func nextResponderFromTextInputView(_ textInputView: some IQTextInputView) -> IQTextInputViewInfoModel? {
+    internal func nextResponderFromTextInputView(_ textInputView: some IQTextInputView) -> IQTextInputViewInfoModel? {
         guard let currentIndex: Int = textInputViewCachedInfoIndex(textInputView),
-                currentIndex < textInputViewInfoCache.count - 1 else { return nil }
+              currentIndex < textInputViewInfoCache.count - 1 else { return nil }
 
         let candidates = Array(textInputViewInfoCache[currentIndex+1..<textInputViewInfoCache.count])
 
@@ -188,19 +182,19 @@ internal extension IQKeyboardReturnKeyHandler {
         }
     }
 
-    func textInputViewCachedInfoIndex(_ textInputView: some IQTextInputView) -> Int? {
+    internal func textInputViewCachedInfoIndex(_ textInputView: some IQTextInputView) -> Int? {
         return textInputViewInfoCache.firstIndex {
             guard let inputView = $0.textInputView else { return false }
             return inputView == textInputView
         }
     }
 
-    func textInputViewCachedInfo(_ textInputView: some IQTextInputView) -> IQTextInputViewInfoModel? {
+    internal func textInputViewCachedInfo(_ textInputView: some IQTextInputView) -> IQTextInputViewInfoModel? {
         guard let index: Int = textInputViewCachedInfoIndex(textInputView) else { return nil }
         return textInputViewInfoCache[index]
     }
 
-    func updateReturnKey(textInputView: some IQTextInputView) {
+    internal func updateReturnKey(textInputView: some IQTextInputView) {
 
         let returnKey: UIReturnKeyType
         if nextResponderFromTextInputView(textInputView) != nil {
@@ -215,11 +209,8 @@ internal extension IQKeyboardReturnKeyHandler {
             textInputView.reloadInputViews()
         }
     }
-}
 
-@available(iOSApplicationExtension, unavailable)
-@MainActor
-extension IQKeyboardReturnKeyHandler {
+    // MARK: Deprecated
 
     @available(*, deprecated, renamed: "lastTextInputViewReturnKeyType")
     @objc public var lastTextFieldReturnKeyType: UIReturnKeyType {
@@ -228,25 +219,521 @@ extension IQKeyboardReturnKeyHandler {
     }
 
     @available(*, deprecated, renamed: "add(textInputView:)")
-    @objc func addTextFieldView(_ textInputView: any IQTextInputView) {
+    @objc public func addTextFieldView(_ textInputView: any IQTextInputView) {
         add(textInputView: textInputView)
     }
 
     @available(*, deprecated, renamed: "remove(textInputView:)")
-    @objc func removeTextFieldView(_ textInputView: any IQTextInputView) {
+    @objc public func removeTextFieldView(_ textInputView: any IQTextInputView) {
         remove(textInputView: textInputView)
     }
 
     @available(*, deprecated, renamed: "addResponderSubviews(of:recursive:)")
-    @objc func addResponderFromView(_ view: UIView, recursive: Bool = true) {
+    @objc public func addResponderFromView(_ view: UIView, recursive: Bool = true) {
         addResponderSubviews(of: view, recursive: recursive)
     }
 
     @available(*, deprecated, renamed: "removeResponderSubviews(of:recursive:)")
-    @objc func removeResponderFromView(_ view: UIView, recursive: Bool = true) {
+    @objc public func removeResponderFromView(_ view: UIView, recursive: Bool = true) {
         removeResponderSubviews(of: view, recursive: recursive)
     }
+
+    // MARK: UITextFieldDelegate
+
+    @objc public func textFieldShouldBeginEditing(_ textField: UITextField) -> Bool {
+
+        var returnValue: Bool = true
+
+        if delegate == nil,
+           let textFieldDelegate: any UITextFieldDelegate = textInputViewCachedInfo(textField)?.textFieldDelegate {
+            if textFieldDelegate.responds(to: #selector((any UITextFieldDelegate).textFieldShouldBeginEditing(_:))) {
+                returnValue = textFieldDelegate.textFieldShouldBeginEditing?(textField) ?? false
+            }
+        }
+
+        if returnValue {
+            updateReturnKey(textInputView: textField)
+        }
+
+        return returnValue
+    }
+
+    @objc public func textFieldShouldEndEditing(_ textField: UITextField) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textFieldDelegate: any UITextFieldDelegate = textInputViewCachedInfo(textField)?.textFieldDelegate {
+            if textFieldDelegate.responds(to: #selector((any UITextFieldDelegate).textFieldShouldEndEditing(_:))) {
+                return textFieldDelegate.textFieldShouldEndEditing?(textField) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @objc public func textFieldDidBeginEditing(_ textField: UITextField) {
+
+        var aDelegate: (any UITextFieldDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textField) {
+                aDelegate = model.textFieldDelegate
+            }
+        }
+
+        aDelegate?.textFieldDidBeginEditing?(textField)
+    }
+
+    @objc public func textFieldDidEndEditing(_ textField: UITextField) {
+
+        var aDelegate: (any UITextFieldDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textField) {
+                aDelegate = model.textFieldDelegate
+            }
+        }
+
+        aDelegate?.textFieldDidEndEditing?(textField)
+    }
+
+    @objc public func textFieldDidEndEditing(_ textField: UITextField, reason: UITextField.DidEndEditingReason) {
+
+        var aDelegate: (any UITextFieldDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textField) {
+                aDelegate = model.textFieldDelegate
+            }
+        }
+
+        aDelegate?.textFieldDidEndEditing?(textField, reason: reason)
+    }
+
+    @objc public func textField(_ textField: UITextField,
+                                shouldChangeCharactersIn range: NSRange,
+                                replacementString string: String) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textFieldDelegate: any UITextFieldDelegate = textInputViewCachedInfo(textField)?.textFieldDelegate {
+            let selector: Selector = #selector((any UITextFieldDelegate).textField(_:shouldChangeCharactersIn:
+                                                                                    replacementString:))
+            if textFieldDelegate.responds(to: selector) {
+                return textFieldDelegate.textField?(textField,
+                                                    shouldChangeCharactersIn: range,
+                                                    replacementString: string) ?? false
+            }
+        }
+        return true
+    }
+
+    @objc public func textFieldShouldClear(_ textField: UITextField) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textFieldDelegate: any UITextFieldDelegate = textInputViewCachedInfo(textField)?.textFieldDelegate {
+            if textFieldDelegate.responds(to: #selector((any UITextFieldDelegate).textFieldShouldClear(_:))) {
+                return textFieldDelegate.textFieldShouldClear?(textField) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @objc public func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        var isReturn: Bool = true
+
+        if let textFieldDelegate: any UITextFieldDelegate = textInputViewCachedInfo(textField)?.textFieldDelegate {
+            if textFieldDelegate.responds(to: #selector((any UITextFieldDelegate).textFieldShouldReturn(_:))) {
+                isReturn = textFieldDelegate.textFieldShouldReturn?(textField) ?? false
+            }
+        }
+
+        if isReturn {
+            goToNextResponderOrResign(from: textField)
+            return true
+        } else {
+            return goToNextResponderOrResign(from: textField)
+        }
+    }
+
+    // MARK: UITextViewDelegate
+
+    @objc public func textViewShouldBeginEditing(_ textView: UITextView) -> Bool {
+
+        var returnValue: Bool = true
+
+        if delegate == nil,
+           let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(textView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector((any UITextViewDelegate).textViewShouldBeginEditing(_:))) {
+                returnValue = textViewDelegate.textViewShouldBeginEditing?(textView) ?? false
+            }
+        }
+
+        if returnValue {
+            updateReturnKey(textInputView: textView)
+        }
+
+        return returnValue
+    }
+
+    @objc public func textViewShouldEndEditing(_ textView: UITextView) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(textView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector((any UITextViewDelegate).textViewShouldEndEditing(_:))) {
+                return textViewDelegate.textViewShouldEndEditing?(textView) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @objc public func textViewDidBeginEditing(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewDidBeginEditing?(textView)
+    }
+
+    @objc public func textViewDidEndEditing(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewDidEndEditing?(textView)
+    }
+
+    @objc public func textView(_ textView: UITextView,
+                               shouldChangeTextIn range: NSRange,
+                               replacementText text: String) -> Bool {
+
+        var shouldChange = true
+
+        if delegate == nil {
+
+            if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(textView)?.textViewDelegate {
+                let selector = #selector((any UITextViewDelegate).textView(_:shouldChangeTextIn:replacementText:))
+                if textViewDelegate.responds(to: selector) {
+                    shouldChange = (textViewDelegate.textView?(textView,
+                                                               shouldChangeTextIn: range,
+                                                               replacementText: text)) ?? true
+                }
+            }
+        }
+
+        if self.dismissTextViewOnReturn, text == "\n" {
+            goToNextResponderOrResign(from: textView)
+            return false
+        }
+
+        return shouldChange
+    }
+
+    @objc public func textViewDidChange(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewDidChange?(textView)
+    }
+
+    @objc public func textViewDidChangeSelection(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewDidChangeSelection?(textView)
+    }
+
+    @available(iOS, deprecated: 17.0)
+    @objc public func textView(_ aTextView: UITextView,
+                               shouldInteractWith URL: URL,
+                               in characterRange: NSRange,
+                               interaction: UITextItemInteraction) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            let selector: Selector = #selector(textView as
+                                               (UITextView, URL, NSRange, UITextItemInteraction) -> Bool)
+            if textViewDelegate.responds(to: selector) {
+                return textViewDelegate.textView?(aTextView,
+                                                  shouldInteractWith: URL,
+                                                  in: characterRange,
+                                                  interaction: interaction) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @available(iOS, deprecated: 17.0)
+    @objc public func textView(_ aTextView: UITextView,
+                               shouldInteractWith textAttachment: NSTextAttachment,
+                               in characterRange: NSRange,
+                               interaction: UITextItemInteraction) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            let selector: Selector = #selector(textView as
+                                               (UITextView, NSTextAttachment, NSRange, UITextItemInteraction)
+                                               -> Bool)
+            if textViewDelegate.responds(to: selector) {
+                return textViewDelegate.textView?(aTextView,
+                                                  shouldInteractWith: textAttachment,
+                                                  in: characterRange,
+                                                  interaction: interaction) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @available(iOS, deprecated: 10.0)
+    @objc public func textView(_ aTextView: UITextView,
+                               shouldInteractWith URL: URL,
+                               in characterRange: NSRange) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector(textView as (UITextView, URL, NSRange) -> Bool)) {
+                return textViewDelegate.textView?(aTextView,
+                                                  shouldInteractWith: URL,
+                                                  in: characterRange) ?? false
+            }
+        }
+
+        return true
+    }
+
+    @available(iOS, deprecated: 10.0)
+    @objc public func textView(_ aTextView: UITextView,
+                               shouldInteractWith textAttachment: NSTextAttachment,
+                               in characterRange: NSRange) -> Bool {
+
+        guard delegate == nil else { return true }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector(textView as (UITextView, NSTextAttachment, NSRange) -> Bool)) {
+                return textViewDelegate.textView?(aTextView,
+                                                  shouldInteractWith: textAttachment,
+                                                  in: characterRange) ?? false
+            }
+        }
+
+        return true
+    }
+
+    // MARK: @available(iOS 16.0, *)
+
+    @available(iOS 16.0, *)
+    public func textView(_ aTextView: UITextView,
+                         editMenuForTextIn range: NSRange,
+                         suggestedActions: [UIMenuElement]) -> UIMenu? {
+
+        guard delegate == nil else { return nil }
+
+        if let textViewDelegate: any UITextViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+
+            let selector: Selector = #selector(textView as
+                                               (UITextView, NSRange, [UIMenuElement]) -> UIMenu?)
+            if textViewDelegate.responds(to: selector) {
+                return textViewDelegate.textView?(aTextView,
+                                                  editMenuForTextIn: range,
+                                                  suggestedActions: suggestedActions)
+            }
+        }
+
+        return nil
+    }
+
+    @available(iOS 16.0, *)
+    public func textView(_ aTextView: UITextView,
+                         willPresentEditMenuWith animator: any UIEditMenuInteractionAnimating) {
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(aTextView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textView?(aTextView, willPresentEditMenuWith: animator)
+    }
+
+    @available(iOS 16.0, *)
+    public func textView(_ aTextView: UITextView,
+                         willDismissEditMenuWith animator: any UIEditMenuInteractionAnimating) {
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(aTextView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textView?(aTextView, willDismissEditMenuWith: animator)
+    }
+
+#if swift(>=5.9)    // Xcode 15
+
+    // MARK: @available(iOS 17.0, *)
+
+    @available(iOS 17.0, *)
+    public func textView(_ aTextView: UITextView,
+                         primaryActionFor textItem: UITextItem,
+                         defaultAction: UIAction) -> UIAction? {
+        guard delegate == nil else { return nil }
+
+        if let textViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector(textView as (UITextView, UITextItem, UIAction) -> UIAction?)) {
+                return textViewDelegate.textView?(aTextView,
+                                                  primaryActionFor: textItem,
+                                                  defaultAction: defaultAction)
+            }
+        }
+
+        return nil
+    }
+
+    @available(iOS 17.0, *)
+    public func textView(_ aTextView: UITextView,
+                         menuConfigurationFor textItem: UITextItem,
+                         defaultMenu: UIMenu) -> UITextItem.MenuConfiguration? {
+        guard delegate == nil else { return nil }
+
+        if let textViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            let selector: Selector = #selector(textView as (UITextView, UITextItem, UIMenu)
+                                               -> UITextItem.MenuConfiguration?)
+            if textViewDelegate.responds(to: selector) {
+                return textViewDelegate.textView?(aTextView,
+                                                  menuConfigurationFor: textItem,
+                                                  defaultMenu: defaultMenu)
+            }
+        }
+
+        return nil
+    }
+
+    @available(iOS 17.0, *)
+    public func textView(_ textView: UITextView,
+                         textItemMenuWillDisplayFor textItem: UITextItem,
+                         animator: any UIContextMenuInteractionAnimating) {
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textView?(textView, textItemMenuWillDisplayFor: textItem, animator: animator)
+    }
+
+    @available(iOS 17.0, *)
+    public func textView(_ textView: UITextView,
+                         textItemMenuWillEndFor textItem: UITextItem,
+                         animator: any UIContextMenuInteractionAnimating) {
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textView?(textView, textItemMenuWillEndFor: textItem, animator: animator)
+    }
+#endif
+
+#if swift(>=6.0)    // Xcode 16
+
+    @available(iOS 18.0, *)
+    @objc public func textViewWritingToolsWillBegin(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewWritingToolsWillBegin?(textView)
+    }
+
+    @available(iOS 18.0, *)
+    @objc public func textViewWritingToolsDidEnd(_ textView: UITextView) {
+
+        var aDelegate: (any UITextViewDelegate)? = delegate
+
+        if aDelegate == nil {
+
+            if let model: IQTextInputViewInfoModel = textInputViewCachedInfo(textView) {
+                aDelegate = model.textViewDelegate
+            }
+        }
+
+        aDelegate?.textViewWritingToolsDidEnd?(textView)
+    }
+
+    @available(iOS 18.0, *)
+    @objc public func textView(_ textView: UITextView,
+                               writingToolsIgnoredRangesInEnclosingRange enclosingRange: NSRange) -> [NSValue] {
+        guard delegate == nil else { return [] }
+
+        if let textViewDelegate = textInputViewCachedInfo(aTextView)?.textViewDelegate {
+            if textViewDelegate.responds(to: #selector(textView as (UITextView, NSRange) -> [NSValue])) {
+                return textViewDelegate.textView?(aTextView,
+                                                  writingToolsIgnoredRangesInEnclosingRange: enclosingRange)
+            }
+        }
+        return []
+    }
+#endif
 }
+// swiftlint:enable type_body_length
 
 @available(iOSApplicationExtension, unavailable)
 @MainActor
@@ -284,3 +771,4 @@ fileprivate extension UIView {
         })
     }
 }
+// swiftlint:enable file_length
