+++ IQKeyboardManagerSwift/IQKeyboardReturnKeyHandler/IQTextInputViewInfoModel.swift
@@ -22,6 +22,7 @@
 //  THE SOFTWARE.
 
 import UIKit
+import IQKeyboardCore
 
 @available(iOSApplicationExtension, unavailable)
 @MainActor
