+++ IQKeyboardManagerSwift/IQKeyboardToolbarManager/IQKeyboardManager+ToolbarManager.swift
@@ -24,6 +24,10 @@
 import UIKit
 
 @available(iOSApplicationExtension, unavailable)
+// swiftlint:disable line_length
+@available(*, deprecated,
+            message: "Please disable Toolbar using `IQKeyboardManager.shared.enableAutoToolbar = false` and use `IQKeyboardToolbarManager` independently from https://github.com/hackiftekhar/IQKeyboardToolbarManager")
+// swiftlint:enable line_length
 @MainActor
 public extension IQKeyboardManager {
