+++ IQKeyboardManagerSwift/IQKeyboardToolbarManager/Toolbar/IQKeyboardToolbarManager+Toolbar.swift
@@ -22,6 +22,7 @@
 //  THE SOFTWARE.
 
 import UIKit
+import IQKeyboardCore
 
 @available(iOSApplicationExtension, unavailable)
 @MainActor
@@ -33,6 +34,7 @@ internal extension IQKeyboardToolbarManager {
     private static let toolbarTag = -1001
 
     // swiftlint:disable function_body_length
+    // swiftlint:disable cyclomatic_complexity
     /**
      Add toolbar if it is required to add on textInputViews and it's siblings.
      */
@@ -124,6 +126,7 @@ internal extension IQKeyboardToolbarManager {
         Self.applyToolbarConfiguration(textInputView: textInputView, toolbarConfiguration: toolbarConfiguration)
     }
     // swiftlint:enable function_body_length
+    // swiftlint:enable cyclomatic_complexity
 
     /** Remove any toolbar if it is IQToolbar. */
     func removeToolbarIfRequired(of textInputView: some IQTextInputView) {    //  (Bug ID: #18)
