+++ IQKeyboardManagerSwift/IQKeyboardToolbarManager/UIKitExtensions/UIView+Responders.swift
@@ -22,6 +22,7 @@
 //  THE SOFTWARE.
 
 import UIKit
+import IQKeyboardCore
 
 @available(iOSApplicationExtension, unavailable)
 @MainActor
@@ -101,7 +102,7 @@ UIView category for managing textInputView
 @available(iOSApplicationExtension, unavailable)
 @MainActor
 internal extension IQKeyboardExtension where Base: UIView {
-    
+
     /**
     Returns all deep subViews of the receiver which canBecomeFirstResponder.
     */
